requirements.txt
flask==3.1.1
mysql-connector-python==9.3.0
flask-wtf==1.2.2
flask-login==0.6.3
email_validator==2.2.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
python-dotenv==1.0.1 